+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/attachmentMiddleware.tsx
@@ -14,71 +14,13 @@ import { ILiveChatWidgetAction } from "../../../../../contexts/common/ILiveChatW
 import { ILiveChatWidgetContext } from "../../../../../contexts/common/ILiveChatWidgetContext";
 import { WebChatActionType } from "../../enums/WebChatActionType";
 import { defaultAttachmentAdaptiveCardStyles } from "./defaultStyles/defaultAtttachmentAdaptiveCardStyles";
-import { defaultAttachmentContentStyles } from "./defaultStyles/defaultAttachmentContentStyles";
-import { defaultAttachmentDividerStyles } from "./defaultStyles/defaultAttachmentDividerStyles";
-import { defaultAttachmentDownloadIconStyles } from "./defaultStyles/defaultAttachmentDownloadIconStyles";
-import { defaultAttachmentFileNameStyles } from "./defaultStyles/defaultAttachmentFileNameStyles";
-import { defaultAttachmentIconStyles } from "./defaultStyles/defaultAtttachmentIconStyles";
 import { defaultAttachmentProps } from "../../../common/defaultProps/defaultAttachmentProps";
-import { defaultAttachmentSizeStyles } from "./defaultStyles/defaultAttachmentSizeStyles";
-import { defaultAttachmentStyles } from "./defaultStyles/defaultAtttachmentStyles";
 import { useChatContextStore } from "../../../../..";
 import { CrossIcon, FileScanInProgressIcon, MaliciousFileIcon } from "../../../../../assets/Icons";
 import { NotificationHandler } from "../../notification/NotificationHandler";
 import { NotificationScenarios } from "../../enums/NotificationScenarios";
 import { defaultMiddlewareLocalizedTexts } from "../../../common/defaultProps/defaultMiddlewareLocalizedTexts";
-
-const AttachmentContent = (props: any) => {
-    return (
-        <div id={props.id} style={props.style} >
-            {props.children}
-        </div>
-    );
-};
-
-const AttachmentIcon = (props: any) => {
-    return (
-        <div id={props.id} style={props.style} >
-            <img src={props.src ?? getFileAttachmentIconData("txt")} />
-        </div>
-    );
-};
-
-const Attachment = (props: any) => {
-    const { iconData, imageCard, textCard, renderer } = props;
-    const [state, ]: [ILiveChatWidgetContext, Dispatch<ILiveChatWidgetAction>] = useChatContextStore();
-    const attachmentId = state.domainStates.renderingMiddlewareProps?.attachmentProps?.webChatAttachmentId ?? defaultAttachmentProps.webChatAttachmentId;
-    const attachmentDividerStyles = {...defaultAttachmentDividerStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentDividerStyles};
-    const attachmentIconStyles = {...defaultAttachmentIconStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentIconStyles};
-    const attachmentStyles = {...defaultAttachmentStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentStyles};
-
-    const attachmentSizeStylesString = (
-        Object.entries({...defaultAttachmentSizeStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentSizeStyles}).map(([k, v]) => `${k.replace(/[A-Z]/g, (match: string) => `-${match.toLowerCase()}`)}:${v}`).join(";"));
-    const attachmentContentStylesString = (
-        Object.entries({...defaultAttachmentContentStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentContentStyles}).map(([k, v]) => `${k.replace(/[A-Z]/g, (match: string) => `-${match.toLowerCase()}`)}:${v}`).join(";"));
-    const attachmentFileNameStylesString = (
-        Object.entries({...defaultAttachmentFileNameStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentFileNameStyles}).map(([k, v]) => `${k.replace(/[A-Z]/g, (match: string) => `-${match.toLowerCase()}`)}:${v}`).join(";"));
-    const attachmentDownloadIconStylesString = (
-        Object.entries({...defaultAttachmentDownloadIconStyles, ...state.domainStates.renderingMiddlewareProps?.attachmentDownloadIconStyles}).map(([k, v]) => `${k.replace(/[A-Z]/g, (match: string) => `-${match.toLowerCase()}`)}:${v}`).join(";"));
-
-    return (
-        <><style>{`
-            .webchat__fileContent__size { ${attachmentSizeStylesString} }
-            .webchat__fileContent { ${attachmentContentStylesString} }
-            .webchat__fileContent__fileName { ${attachmentFileNameStylesString} }
-            .webchat__fileContent__downloadIcon { ${attachmentDownloadIconStylesString} }
-        `}</style>
-        <div dir={state.domainStates.globalDir}>
-            {imageCard && renderer(imageCard)}
-            {imageCard && <hr id={attachmentId + "-divider"} style={attachmentDividerStyles} />}
-            <AttachmentContent attachment={textCard.attachment} id={attachmentId} style={attachmentStyles}>
-                <AttachmentIcon src={iconData} id={attachmentId + "-icon"} style={attachmentIconStyles} />
-                {textCard && renderer(textCard)}
-            </AttachmentContent>
-        </div>
-        </>
-    );
-};
+import Attachment from "./attachments/Attachment";
 
 const Spinner = (props: any) => {
     const spinnerStyle: any = {
