+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/attachments/AttachmentContent.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const AttachmentContent = (props: any) => { // eslint-disable-line @typescript-eslint/no-explicit-any
+    return (
+        <div id={props.id} style={props.style} >
+            {props.children}
+        </div>
+    );
+};
+
+export default AttachmentContent;
